Add editable dev server URL, login persistence, and fix mock data
- Make dev server URL editable on login screen instead of static text
- Persist username and dev URL across sessions via DataStore
- Fix mock data not sending when using custom dev URL by checking
  against production URL instead of default URL
- Change SpO2 DTO field from Double to Int for clean serialization
- Round HRV mock values to 2 decimal places
- Allow cleartext to any host in network_security_config.xml
- Add README.md

+++ app/src/main/kotlin/com/gymdash/companion/data/mapper/HealthDataMapper.kt
@@ -178,7 +178,7 @@ class HealthDataMapper @Inject constructor() {
                     spO2Readings.add(
                         SpO2ReadingSync(
                             timestamp = record.time.toString(),
-                            spO2Percentage = record.percentage.value
+                            spO2Percentage = record.percentage.value.toInt()
                         )
                     )
                 }
